% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customized_compute.R
\name{sc.metabolism.customized}
\alias{sc.metabolism.customized}
\title{scMetabolismplus}
\usage{
sc.metabolism.customized(
  obj,
  method = "AUCell",
  input_pathway,
  imputation = F,
  ncores = 20,
  input_dir
)
}
\arguments{
\item{obj}{seruat 对象}

\item{input_pathway}{#输入的通路和所属类别
使用的核心数量，默认为 20}
}
\description{
scMetabolism
}
\examples{
sc.metabolism.customized()
}
\keyword{scMetabolismplus}
