% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways_Seurat.R
\name{sc.metabolism.Seurat.pathway}
\alias{sc.metabolism.Seurat.pathway}
\title{scMetabolismplus}
\usage{
sc.metabolism.Seurat.pathway(
  obj,
  method = "AUCell",
  imputation = F,
  Cancer = "BRCA",
  metabolism.type = "KEGG",
  ncore = 20
)
}
\arguments{
\item{obj}{seruat 对象}

\item{Cancer}{数据属于的癌种 对象}
}
\description{
scMetabolism
}
\examples{
sc.metabolism.Seurat.pathway()
}
\keyword{scMetabolismplus}
