% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaolism_Seurat.R
\name{sc.metabolism.Seurat}
\alias{sc.metabolism.Seurat}
\title{scMetabolismplus}
\usage{
sc.metabolism.Seurat(
  obj,
  method = "AUCell",
  imputation = F,
  Cancer = "BRCA",
  metabolism.type = "KEGG"
)
}
\arguments{
\item{obj}{seruat 对象}

\item{Cancer}{数据属于的癌种 对象}
}
\description{
scMetabolism
}
\examples{
sc.metabolism.Seurat()
}
\keyword{scMetabolismplus}
